Happy with this PR yet?  Well I'm not...
... Let's make the sample rate in the options in the features in the LV2Wrapper
in the LV2Instance -- immutable.

We decided earlier that we won't support changing of rates of plug-in instances,
and there may be plug-ins for which that is impossible anyway.

+++ src/effects/lv2/LV2Wrapper.h
@@ -60,7 +60,7 @@ public:
    //! Constructor may spawn a thread
    LV2Wrapper(CreateToken&&,
       const LV2FeaturesList &featuresList,
-      const LilvPlugin &plugin, double sampleRate);
+      const LilvPlugin &plugin, float sampleRate);
 
    //! If a thread was started, joins it
    ~LV2Wrapper();
